requirements.txt
nextcord==2.3.2
python-dotenv==0.21.1
requests==2.28.2
configparser==5.3.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
websockets==10.4
SQLAlchemy==2.0.3